scripts/pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "biopython>=1.86",
    "matplotlib>=3.10.8",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pycountry-convert>=0.7.2",
    "scikit-learn>=1.8.0",
]
